+++ server/schema/ql-js-generator.js
@@ -459,7 +459,7 @@ function main(schemaDef: TypeDef) {
         `);
 
         types.forEach((type: DbType) => {
-            ql.writeLn(`\t${type.collection || ''}(filter: ${type.name}Filter, orderBy: [QueryOrderBy], limit: Int, timeout: Float, auth: String): [${type.name}]`);
+            ql.writeLn(`\t${type.collection || ''}(filter: ${type.name}Filter, orderBy: [QueryOrderBy], limit: Int, timeout: Float, accessKey: String): [${type.name}]`);
         });
 
         ql.writeBlockLn(`